Still playing with effects

+++ src/main/java/com/matthewcasperson/utils/GeneralUtils.java
@@ -5,11 +5,11 @@ public interface GeneralUtils {
      * A no-throw sleep method
      * @param milliseconds How long to sleep for
      */
-    void sleep(final int milliseconds);
+    void sleep(final long milliseconds);
     /**
      * A no-throw sleep method
      * @param milliseconds How long to sleep for
      * @param nanoseconds Nanoseconds to sleep for
      */
-    void sleep(final int milliseconds, final int nanoseconds);
+    void sleep(final long milliseconds, final int nanoseconds);
 }
